api/api/requirements.txt
# ✅ Render.com deployment requirements - All dependencies flattened
# CRITICAL: databases package for Manager Agent system
databases[postgresql]>=0.7.0
asyncpg>=0.29.0

# ── Core FastAPI runtime ────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn>=0.34.0
python-multipart>=0.0.6  # Required for FastAPI File/UploadFile (form-data)
httpx>=0.27.0
pydantic>=2.10,<3
python-dotenv>=0.20.0,<1
requests>=2.0,<3
typing-extensions>=4.12.2,<5

# ── Supabase integration ──────────────────────────────────────────────
# Using PyPI version instead of git for Render reliability
supabase>=2.1.0,<3.0.0

# ── Validation / schema ────────────────────────────────────────────
jsonschema>=4.21

# ── Development / testing (Render needs these) ─────────────────────────
pytest
pytest-asyncio
pytest-mock
ruff
pre-commit

# ── LLM integrations (using PyPI versions for stability) ──────────────
openai>=1.66.5
anthropic>=0.40.0  # Required for Claude Files API download in work_outputs

# ── Documentation tooling ─────────────────────────────────────────────  
griffe>=1.5.6,<2
mkdocs-static-i18n>=1.3.0
types-requests>=2.0,<3

# ── MCP (Model Context Protocol) ──────────────────────────────────────
mcp>=1.6.0,<2; python_version >= "3.10"

# ── File Processing (Memory-First approach) ───────────────────────────
pymupdf>=1.24.0
httpx>=0.27.0  # For downloading Supabase Storage files

# ── OCR Processing (optional for direct file bytes) ───────────────────
# Pillow (PIL) for image processing
Pillow>=10.0.0
# pytesseract for OCR - requires tesseract system package
pytesseract>=0.3.10

# ── Telegram Bot ──────────────────────────────────────────────────────
python-telegram-bot>=21.0

# ── Scheduling ────────────────────────────────────────────────────────
APScheduler>=3.10.0
